rename: utils 안에 상수, 타입 폴더 삽입
